requirements.txt
django
djangorestframework
pyjwt
drf-yasg